Remove multi_json dependency by requiring Ruby 1.9
This was the last lingering dependency. I was able to convert almost
every method to request and parse YAML instead of JSON but I got stuck
on rubygems/gemcutter#323. Ruby 1.9 includes JSON as part of the
standard library, so by requiring 1.9, I don't need to specify it as a
dependency. Users of older Ruby versions can hack around this constraint
by force-installing the gem and requiring a JSON library manually.

Closes #4.