Add very basic error support to Gems
Rather than returning a string when an error occurs, if a request is
returned with something other than a 200 or a redirection, we raise
an error with the response body as the message.

This pattern can be expanded upon to add more specific error handling
for more specific cases (such as NotFound or Unauthorized).

Signed-off-by: Tom Duffield <tom@chef.io>